requirements.txt
langchain
langchain[openai]
python-dotenv
uuid
requests